simplify ChatList
