+++ frontend/app/layout.tsx
@@ -56,9 +56,7 @@ export default function RootLayout({ children }: RootLayoutProps) {
           disableTransitionOnChange
         >
           <div className="flex h-dvh min-h-0 flex-col">
-            <main className="flex min-h-0 flex-1 flex-col bg-muted/50">
-              {children}
-            </main>
+            <main className="flex min-h-0 flex-1 flex-col">{children}</main>
           </div>
         </Providers>
       </body>
