+++ frontend/components/sidebar-desktop.tsx
@@ -3,7 +3,7 @@ import { Sidebar } from './sidebar'
 import { ChatHistory } from './chat-history'
 export async function SidebarDesktop() {
   return (
-    <Sidebar className="relative inset-y-0 z-30 hidden -translate-x-full border-r bg-muted bg-surface-0 duration-300 ease-in-out data-[state=open]:translate-x-0 md:flex md:w-[220px] lg:w-[300px]">
+    <Sidebar className="relative inset-y-0 z-30 hidden -translate-x-full border-r bg-surface-0 duration-300 ease-in-out data-[state=open]:translate-x-0 md:flex md:w-[220px] lg:w-[300px]">
       <ChatHistory
         mobile={false}
         logoLockup={<LogoLockup className="h-[26px] w-auto" />}
